Move stray ApiClient functions to utils modules
