+++ src/utils/strings.js
@@ -1,0 +1,32 @@
+/**
+ * Utility module for string operations
+ * @module utils/strings
+ */
+
+export function replaceAll(originalString, strReplace, strWith) {
+    const reg = new RegExp(strReplace, 'ig');
+    return originalString.replace(reg, strWith);
+}
+
+export function compareVersions(a, b) {
+    // -1 a is smaller
+    // 1 a is larger
+    // 0 equal
+    a = a.split('.');
+    b = b.split('.');
+
+    for (let i = 0, length = Math.max(a.length, b.length); i < length; i++) {
+        const aVal = parseInt(a[i] || '0');
+        const bVal = parseInt(b[i] || '0');
+
+        if (aVal < bVal) {
+            return -1;
+        }
+
+        if (aVal > bVal) {
+            return 1;
+        }
+    }
+
+    return 0;
+}
